federated-module/fl_scripts/requirements.txt
numpy==1.24.3
kafka-python==2.2.15
tensorflow-macos
tensorflow-io==0.36.0
